[jax2tf] Improve handling of XlaSharding ops.
Since TF does not support the XlaSharding op in eager mode, we make two changes: raise an error when a function with sharded arguments or results is used in a TF eager context, and omit the XlaSharding op for REPLICATED sharding. The latter improves usability, allowing lowered functions with replicated sharding to be used in TF eager context. It also results in smaller graphs.

PiperOrigin-RevId: 484174652